abstract base client
